Code Quality:
* Separate a bunch of TrainManager code out and comment
