+++ openBVE/OpenBve/OpenBve.csproj
@@ -130,8 +130,11 @@
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
     <Compile Include="Simulation\TrainManager\Doors.cs" />
+    <Compile Include="Simulation\TrainManager\Train\BrakeSystem.cs" />
+    <Compile Include="Simulation\TrainManager\Train\Handles.cs" />
     <Compile Include="Simulation\TrainManager\Passengers.cs" />
     <Compile Include="Simulation\TrainManager\Station.cs" />
+    <Compile Include="Simulation\TrainManager\Train\Reverser.cs" />
     <Compile Include="Simulation\World\Background.cs" />
     <Compile Include="Simulation\World\Transformations.cs" />
     <Compile Include="System\Functions\CrashHandler.cs" />
